requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
markdownify==0.11.6
google-api-python-client==2.118.0
google-auth==2.28.0
pytest==8.0.0
pytest-mock==3.12.0
responses==0.25.0
